chore: move launchers to separate file
